+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added TableModel support in JCommand
+
 ## [2.3.2] - 2024-06-13
 
 ### Changed